% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{assign,character,ANY,object_store-method}
\alias{assign,character,ANY,object_store-method}
\title{Upload an R Value to an Object Store}
\usage{
\S4method{assign}{character,ANY,object_store}(x, value, pos)
}
\arguments{
\item{x}{A key name, optionally including a \code{/} separated directory path}

\item{value}{Any serializeable R value.}

\item{pos}{An object store connection from \code{\link{connect}}.}
}
\value{
A character string corresponding to the url of the uploaded object.
}
\description{
Upload the R \code{value} to the object store connection \code{pos} with the key name and
optional path specified by \code{x}. This method is analagous to the standard R
\code{assign()} function.
}
\note{
Key names are url-encoded and may be changed (\code{assign()} returns the uri of the
stored value). The forward slash character \code{/} is NOT url-encoded and reserved for directory
path information. Do not use any slash (forward or backward) in your key names, they will
be interpreted as directory separators.
}
\examples{
# Start an example local mongoose backend server
mongoose_start(path=tempdir())
con <- connect()
# Store the 'iris' dataset in a directory named 'mydata':
assign("mydata/iris", iris, con)
# Retrieve it from the object store into a new variable called 'x'
x <- get("mydata/iris", con)
# Delete the entire 'mydata' directory
delete("mydata", con)
mongoose_stop()
}
\seealso{
\code{\link{connect}} \code{\link{get,character,object_store-method}} \code{\link{remove}}
}

