% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNinst.R
\name{checkNinst}
\alias{checkNinst}
\title{Check the presence of and install packages}
\usage{
checkNinst(
  pk = NULL,
  repos = "https://cloud.r-project.org/",
  inst = FALSE,
  ...
)
}
\arguments{
\item{pk}{a character vector with the names of the packages to check}

\item{repos}{(optional) the URL to a CRAN repository. Default to https://cloud.r-project.org. Get URLs here: https://cran.r-project.org/mirrors.html}

\item{inst}{logical. Should missing packages be installed? Default to FALSE.}

\item{...}{extra arguments to \link{install.packages}}
}
\description{
The function verify if required packages for the course are installed and, if not, install them.
Used at the beginning of a project could help retriving the missing packages.
}
\details{
The user is asked if the missing packages should be installed.
}
\seealso{
\link{install.packages}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
