% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{connect}
\alias{connect}
\title{Register an Object Store Service}
\usage{
connect(url = "http://localhost:8000", backend = mongoose, ...)
}
\arguments{
\item{url}{The service root url including protocol, address and port number. For example \code{http://localhost:8000}.}

\item{backend}{A service backend provider function. The default is \code{mongoose}, but you may choose from other available
backends like \code{minio} and {s3}.}

\item{...}{Backend-specific arguments, see backend documentation for details.}
}
\value{
An object of class \code{object_store}, really a function usable by \code{assign}, \code{get}, \code{delete}, and \code{info}.
}
\description{
Register an object store service backend, including backend-specific options
like authentication, encryption, and compression.
}
\examples{
# Start an example local mongoose backend server
mongoose_start(path=tempdir())
con <- connect()
# Store part of the 'iris' dataset in a directory named 'mydata':
assign("mydata/iris", head(iris), con)
# Retrieve it from the object store:
get("mydata/iris", con)
# Retrieve the URL of the object:
get("mydata/urus", con, mode="url")
# Delete the entire 'mydata' directory
delete("mydata", con)
mongoose_stop()
}
\seealso{
\code{\link{mongoose}} \code{link{get,character,object_store-method}} \code{\link{assign,character,ANY,object_store-method}} \code{\link{delete}}
}
