% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{delete}
\alias{delete}
\title{Deleta a Value or Directory from an Object Store}
\usage{
delete(key, con)
}
\arguments{
\item{key}{A key name, optionally including a \code{/} separated directory path}

\item{con}{An object store connection from \code{\link{connect}}.}
}
\value{
\code{NULL} is invisibly returned, or an error may be thrown.
}
\description{
Delete the value or directory corresponding to the specified \code{key} from the object store service
connection \code{con}.
This function roughly corresponds to R's \code{remove()} function.
}
\examples{
# Start an example local mongoose backend server
mongoose_start(path=tempdir())
con <- connect()
# Store the 'iris' dataset in a directory named 'mydata':
assign("mydata/iris", iris, con)
# Retrieve it from the object store into a new variable called 'x'
x <- get("mydata/iris", con)
# Delete the entire 'mydata' directory
delete("mydata", con)
mongoose_stop()
}
\seealso{
\code{\link{connect}} \code{\link{assign,character,ANY,object_store-method}} \code{\link{get,character,object_store-method}}
}

