% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_dictionary.R
\name{export_dictionary}
\alias{export_dictionary}
\title{Export data dictionary}
\usage{
export_dictionary(
  description = NULL,
  varTable = NULL,
  path = NULL,
  formatTable = "table",
  overwrite = TRUE
)
}
\arguments{
\item{description}{character containing the info additional to the data dictionary}

\item{varTable}{a data.frame object containing description and metadata of the variables in the related dataset. Typically the dataframe is initated using \link{initiate_dictionary}.}

\item{path}{a character specifying the path and the name of the file to be saved.}

\item{formatTable}{character. Only considered for .txt output. One of the following two options: 'long', 'table'(DEFAULT). 'long' format will report one variable and its metadata at the time, one after the other. 'table' format will paste metadata as a formatted table.}

\item{overwrite}{logical. should overwrite an existing file with the same name? DEFAULT to TRUE. If FALSE than the file is appended.}
}
\value{
Export the data dictionary file as a .txt or .csv (see Details), or formats that can be saved by \link[utils:write.table]{utils::write.table}.
}
\description{
Export the data dictionary for a dataset with possibility to add additional information (e.g. dataset description, reference system for geographical coordinates, fiel creation date,... ).
The data dictionary is formatted based on the CHONe dataverse guideline at \url{https://chone2.ca/app/uploads/2020/05/CHONe-Dataverse-User-Guideline_v3.pdf}.
Additional information can be added before the data dictionary.
}
\details{
Consider the following format depending on the situation:
\itemize{
\item ".txt" If exporting both dictionary (i.e. varTable) and additional information (i.e. description). This format is more easily human readable.
\item ".csv" If exporting only the dictionary (i.e. varTable). If description is present, varTable is appended after
}

When format is ".csv" dictionary is saved by write.table with row.names=F, col.names=TRUE, sep=',', dec='.', quote=T
}
\examples{
# initiate a data dictionary
dict_names<-c('field_name','description', 'data_type', 'unit','missing_data_code')
dict<- matrix(nrow=5, ncol=length(dict_names), NA)
dict<-as.data.frame(dict)
names(dict)<-dict_names
# fill in a description of the dataset 
dict[,'field_name']<-letters[1:5]
dict[,'description']<- paste0('this is variable ',LETTERS[1:5])


# example of additional info
v1<-12345
daytag<-as.Date(Sys.time())
txt<-paste0('R SCRIPT 01-urchin_rocks-data_extraction on ',daytag,'
DATASET DESCRIPTION
Dataset describing the substrate on which Green Sea Urchin (gsu) where found in Bay Sept-Iles in July 2018
Coordinate Reference System
EPSG: ',v1,'
proj4string: ',v1,'
'
)

file_name<-paste0(tempfile(),'.txt')
export_dictionary(description=txt,varTable=dict, path=file_name)
# open exported distionary in the text editor
edit(file=file_name)

file_name<-paste0(tempfile(),'.csv')
export_dictionary(description=txt,varTable=dict, path=file_name)
# open exported distionary in the text editor
edit(file=file_name)


}
\seealso{
\link{initiate_dictionary}
}
\author{
Filippo Ferrario, \email{filippo.f3rrario@gmail.com}
}
