% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{get,character,object_store,ANY,ANY-method}
\alias{get,character,object_store,ANY,ANY-method}
\title{Retrieve a Value from an Object Store}
\usage{
\S4method{get}{character,object_store,ANY,ANY}(x, pos)
}
\arguments{
\item{x}{A key name, optionally including a \code{/} separated directory path.}

\item{pos}{An object store connection from \code{\link{connect}}.}

\item{mode}{Optional, set \code{mode="url"} to return the URL of the specified object.}
}
\value{
Either a data frame directory listing when \code{x} corresponds to a directory,
or an R value corresponding to \code{x}.
}
\description{
Retrieve a value corresponding to the specified name \code{x} from the object store service
connection \code{pos}. The method is analagous to the standard R \code{get()} function.
If \code{x} corresponds to a directory path, then a data frame listing
the directory contents is returned. Set \code{x=""} to list the contents of the
service root directory path.
}
\note{
Directory entries in the data frame directory listing output are identified by \code{size=NA}.
}
\examples{
# Start an example local mongoose backend server
mongoose_start(path=tempdir())
con <- connect()
# Cache the 'iris' dataset in a directory named 'mydata':
assign("mydata/iris", head(iris), con)
# Retrieve it from the object store
get("mydata/iris", con)
# Retrieve only the (character) url of the object:
get("mydata/iris", con, mode="url")
# Delete the entire 'mydata' directory
delete("mydata", con)
mongoose_stop()
}
\seealso{
\code{\link{connect}} \code{\link{assign,character,ANY,object_store-method}} \code{\link{remove}}
}

