% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{info}
\alias{info}
\title{Retrieve Metadata from an Object Store}
\usage{
info(key, con)
}
\arguments{
\item{key}{A key name, optionally including a \code{/} separated directory path.}

\item{con}{An object store connection from \code{\link{connect}}.}
}
\value{
Either a data frame directory listing when \code{key} corresponds to a directory,
or an R list of headers and their values corresponding to \code{key}.
}
\description{
Retrieve metadata like last modified time and size corresponding to the specified \code{key} from the object store service
connection \code{con}. If \code{key} corresponds to a directory, then a data frame listing
the directory contents is returned.
}
\note{
Corresponds to the HTTP \code{HEAD} operations. Directory entries in the data frame directory listing output are identified by \code{size=NA}.
}
\examples{
# Start an example local mongoose backend server
mongoose_start(path=tempdir())
con <- connect()
# Store the 'iris' dataset in a directory named 'mydata':
assign("mydata/iris", iris, con)
# Print some info about it
info("mydata/iris", con)
# Retrieve it from the object store into a new variable called 'x'
x <- get("mydata/iris", con)
# Delete the entire 'mydata' directory
delete("mydata", con)
mongoose_stop()
}
\seealso{
\code{\link{connect}} \code{\link{get}} \code{\link{delete}}
}

