% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{initiate_dictionary}
\alias{initiate_dictionary}
\title{Initiate data dictionary}
\usage{
initiate_dictionary(dataset)
}
\arguments{
\item{dataset}{a data.frame object to be documented.}
}
\value{
The data.frame of the dictionary to be completed and saved along with the dataset.
}
\description{
Initiate the data dictionary to document a dataset.
The data dictionary is formatted based on the CHONe dataverse guideline at \url{https://chone2.ca/app/uploads/2020/05/CHONe-Dataverse-User-Guideline_v3.pdf}.
}
\examples{

# create a data.frame
mydata<-data.frame(ind=LETTERS[1:5],score=1:5)
# initiate the dictionary 
dict_mydata<-initiate_dictionary(mydata)

dict_mydata

# fill in the dictionary
dict_mydata[dict_mydata$field_name=='ind',2:4]<-c('Individual ID','string','N/A')
dict_mydata[dict_mydata$field_name=='score',2:4]<-c('score of Ind','numeric','apples')

dict_mydata

}
\seealso{
\link{export_dictionary}
}
\author{
Filippo Ferrario, \email{filippo.ferrario.1@ulaval.ca}
}
