% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{ls,object_store-method}
\alias{ls,object_store-method}
\title{List the contents of an \code{object_stor} connection object}
\usage{
\S4method{ls}{object_store}(name)
}
\arguments{
\item{name}{\code{object_store} object}

\item{...}{optiona arguments}
}
\description{
List the contents of an \code{object_stor} connection object
}

