% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnvar.R
\docType{data}
\name{mnvar}
\alias{mnvar}
\title{Mean and variances of sources for generating lognormal data}
\format{A matrix with 2 rows and 5 variables}
\description{
Note: on the log scale
}
\details{
\itemize{
		\item source1. mean, variance of first source
    	\item source2. mean, variance of first source
     \item source3. mean, variance of first source
     \item source4. mean, variance of first source
     \item source5. mean, variance of first source
	...
}
}
\keyword{datasets}
