% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoose.R
\name{mongoose_start}
\alias{mongoose_start}
\title{Start a Mongoose Service
Manuall start a local mongoose file-backed storage service}
\usage{
mongoose_start(path = getwd(), port = 8000L, forward_to = NULL,
  ssl_cert = NULL, auth_domain = NULL, global_auth = NULL)
}
\arguments{
\item{path}{full path to data directory, defaults to the current working directory}

\item{port}{service port number}

\item{forward_to}{forward 'not found' requests to another server}

\item{ssl_cert}{optional SSL certificate for TLS-encrypted communication (if you specify this, the mongoose server will only use TLS encyption; otherwise no encryption is used)}

\item{auth_domain}{HTTP digest authentication domain/realm}

\item{global_auth}{HTTP digest global authentication file (with full path)}
}
\value{
Nothing; the mongoose server is started up as a background process.
}
\description{
Start a Mongoose Service
Manuall start a local mongoose file-backed storage service
}
\note{
Leave parameters \code{NULL} to not use the corresponding features.
}
\seealso{
\code{\link{htdigest}}
}

