% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitorsNE.R
\docType{data}
\name{monitorsNE}
\alias{monitorsNE}
\title{Locations of 85 northeast US monitors used in SHARE study}
\format{A dataframe with 85 rows and 3 variables}
\description{
\itemize{
		\item Latitude. Latitude in degrees (decimal form)
    	\item Longitude. Longitude in degrees (decimal form)
     \item monitor. Monitor name to match with constituent concentrations (consConc.rda)
	...
}
}
\keyword{datasets}

