\name{nameSeq-methods}
\docType{methods}
\alias{nameSeq}
\alias{nameSeq-methods}
\alias{nameSeq,genoSet-method}
\alias{nameSeq,haploSet-method}
\title{Name of Sequences}
\description{
  Methods to return the names of the sequences in the object.
}
\section{Methods}{
\describe{
\item{.Object = "genoSet"}{Names of the genotype sequences in the
  \code{genoSet} object}
\item{.Object = "haploSet"}{Names of the haplotype sequences in the
  \code{haploSet} object}
}}
\keyword{methods}
