\name{nameSNP-methods}
\docType{methods}
\alias{nameSNP}
\alias{nameSNP-methods}
\alias{nameSNP,genoSet-method}
\alias{nameSNP,haploSet-method}
\title{Name of SNPs}
\description{
  Methods to return the names of the SNPs in the object.
}
\section{Methods}{
\describe{
\item{.Object = "genoSet"}{Names of the SNPs in the genotype sequences in the
  \code{genoSet} object}
\item{.Object = "haploSet"}{Names of the SNPs in the haplotype sequences in the
  \code{haploSet} object}
}}
\keyword{methods}
