\name{nSeq-methods}
\docType{methods}
\alias{nSeq}
\alias{nSeq-methods}
\alias{nSeq,genoSet-method}
\alias{nSeq,haploSet-method}
\title{Number of Sequences}
\description{
  Methods to count how many sequences in the object.
}
\section{Methods}{
\describe{
\item{.Object = "genoSet"}{Number of genotype sequences in the
  \code{genoSet} object}
\item{.Object = "haploSet"}{Number of haplotype sequences in the
  \code{haploSet} object}
}}
\keyword{methods}
