\name{nSNP-methods}
\docType{methods}
\alias{nSNP}
\alias{nSNP-methods}
\alias{nSNP,genoSet-method}
\alias{nSNP,haploSet-method}
\title{Number of SNPs}
\description{
  Methods to count how many SNPs in the object.
}
\section{Methods}{
\describe{
\item{.Object = "genoSet"}{Number of SNPs in the
  \code{genoSet} object}
\item{.Object = "haploSet"}{Number of SNPs in the
  \code{haploSet} object}
}}
\keyword{methods}
