% Generated by roxygen2 (4.0.1): do not edit by hand
\name{outervPCA}
\alias{outervPCA}
\title{Find regional sources}
\usage{
outervPCA(data, ndays = 50, cut = 1, nmsource = NULL)
}
\arguments{
\item{data}{data frame of daily constituent concentrations with date as first column}

\item{ndays}{number of days of data necessary to apply PCA}

\item{nmsource}{number of major sources.  If null, uses number of eigenvalues of the correlation matrix greater than cut.}

\item{cut}{cutoff for eigenvalues (see nmsource), default is 1.}
}
\description{
\code{outervPCA} Applies vPCA locally and across all monitors
}
\details{
Applies vPCA to data from each monitor, and then
a second vPCA to the PCs from the first stage to find regional sources
}

