% Generated by roxygen2 (4.0.2): do not edit by hand
\name{percinc}
\alias{percinc}
\title{\code{percinc} Percent increase}
\usage{
percinc(beta, scale = 10)
}
\arguments{
\item{beta}{regression coefficients from poisson model}

\item{scale}{amount for percent increase}
}
\description{
\code{percinc} Percent increase
}

