% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{print,object_store-method}
\alias{print,object_store-method}
\title{Print a summary of an \code{object_store} object}
\usage{
\S4method{print}{object_store}(x)
}
\arguments{
\item{x}{\code{object_store} object}
}
\value{
printed object summary
}
\description{
Print a summary of an \code{object_store} object
}

