% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.object_store}
\alias{print.object_store}
\title{Print a summary of an \code{object_store} connection object}
\usage{
\method{print}{object_store}(x, ...)
}
\arguments{
\item{x}{\code{object_store} object}

\item{...}{optional arguments (not used)}
}
\value{
printed object summary
}
\description{
Print a summary of an \code{object_store} connection object
}
