% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share-package.r, R/share.R
\docType{package}
\name{share}
\alias{share}
\alias{share-package}
\title{share.}
\usage{
share(data, cut = 1, nmsources = NULL, thres = pi/4, ndays = 50)
}
\arguments{
\item{data}{data frame of daily constituent concentrations with date as first column}

\item{cut}{cutoff for eigenvalues (see nmsource), default is 1.}

\item{nmsources}{number of major sources.  If null, uses number of eigenvalues of the correlation matrix greater than cut.}

\item{thres}{cutoff for angle between local and major sources}

\item{ndays}{number of days of data necessary to apply PCA for each monitor}
}
\description{
share.

\code{share} estimates major sources of pollution using the share method
}
\details{
This function estimates major sources of pollution across
multiple ambient monitors.  Other outputs of share include
a matrix to guide pooling short-term health effects of pollution
to conduct regional and national studies.
Works on list where each element of the list is a 
dataframe corresponding to one ambient monitor.  
The first column of each dataframe is date and all subsequent 
columns are concentrations of chemical constituents.
}
\examples{
data(consConc)
share(consConc)
}

