% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sharehealth}
\alias{sharehealth}
\title{Find sources SHared Across a REgion (SHARE)}
\usage{
sharehealth(consdata, healthdata = NULL, list = NULL, formula = NULL,
  lag = NULL, groupvar = NULL, print = F, cut = 1, thres = pi/4,
  tots = NULL, method = "SHARE")
}
\arguments{
\item{consdata}{list of data frames of daily constituent concentrations with date as first column for each monitor}

\item{healthdata}{list of data frames corresponding to health counts and covariates with date as first column for each community}

\item{list}{list of length equal to healthdata indicating which monitors correspond to which elements in healthdata}

\item{tots}{list of vectors corresponding to total concentrations (total PM2.5) for each day and monitor.  If null, uses \code{rowSums(data)}}

\item{method}{regional source apportionment method (SHARE or mAPCA)}
}
\description{
\code{sharehealth} Estimates regional health effects using SHARE
}
\details{
This function 1) Applies SHARE, 2) Estimates sources using
APCA at each monitor, 3) Estimates health effects for each community,
and 4) Estimates regional health effects for each source
}

