% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,object_store-method}
\alias{show,object_store-method}
\title{Print a summary of a \code{object_store} object}
\usage{
\S4method{show}{object_store}(object)
}
\arguments{
\item{object}{an object of class \code{object_store}}
}
\value{
printed object summary
}
\description{
Print a summary of a \code{object_store} object
}

