% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlnise2.R
\name{tlniseC}
\alias{tlniseC}
\title{share}
\usage{
tlniseC(Y, V, w = NA, V0 = NA, prior = NA, N = 1000, seed = NULL,
  Tol = 1e-06, maxiter = 1000, intercept = TRUE, labelY = NA,
  labelYj = NA, labelw = NA, digits = 4, brief = 1, prnt = TRUE)
}
\description{
\code{tlniseC} tlnise function from tlnise package
}
\details{
Reproduces tlnise function from tlnise package, but gives convergence output.
See ?tlnise for more info.
}
