% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{toenv}
\alias{toenv}
\title{Convert an object store connection object to an environment of promises
Convert an object store connection object to an environment of promises.
Objects are lazily retrieved from the object store using \code{\link{get()}} when accessed.}
\usage{
toenv(x, base = "")
}
\arguments{
\item{x}{an object store, see \code{connect()}, \code{\link{setAs}}}

\item{base}{base path}
}
\value{
an environment
}
\description{
Convert an object store connection object to an environment of promises
Convert an object store connection object to an environment of promises.
Objects are lazily retrieved from the object store using \code{\link{get()}} when accessed.
}
\keyword{function}
\keyword{internal}

