% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vPCAf}
\alias{vPCAf}
\title{\code{vPCAf} Applies PCA and vPCA and formats outcome}
\usage{
vPCAf(data, i, nsources = NULL, cut = 1)
}
\arguments{
\item{data}{data frame of daily constituent concentrations}

\item{nsource}{number of sources.  If null, uses number of eigenvalues of the correlation matrix greater than cut.}

\item{cut}{cutoff for eigenvalues (see nsource), default is 1.}
}
\description{
\code{vPCAf} Applies PCA and vPCA and formats outcome
}

