% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{whichCS}
\alias{whichCS}
\title{\code{whichCS} Find matches from matchfun}
\usage{
whichCS(locreg)
}
\arguments{
\item{locreg}{matrix of nsources by nmsources of matches}
}
\description{
\code{whichCS} Find matches from matchfun
}

