% Generated by roxygen2 (4.0.1): do not edit by hand
\name{angle}
\alias{angle}
\title{\code{angle} Computes angle between major and local source signatures}
\usage{
angle(major.sig, source.sig1)
}
\arguments{
\item{major.sig}{major source signatures}

\item{source.sig1}{local source signature from one monitor}
}
\description{
\code{angle} Computes angle between major and local source signatures
}
\examples{
data(simdat)
share1 <- share(simdat)
angle(share1$major.sig, share1$source.sig)
}

