% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mAPCA}
\alias{mAPCA}
\title{Functions for mAPCA}
\usage{
mAPCA(data, tots = NULL, nsources = NULL, adjust = NULL, mdl = NULL,
  cut = 1, lim = 50)
}
\arguments{
\item{data}{list of data frames of daily constituent concentrations with date as first column}

\item{tots}{list of totals (PM2.5) corresponding to each monitor}

\item{nsource}{number of sources.  If null, uses number of eigenvalues of the correlation matrix greater than cut.}

\item{adjust}{method of handles adjustment}

\item{mdl}{list of MDLs for each monitor if applicable}

\item{cut}{cutoff for eigenvalues (see nsource), default is 1.}

\item{lim}{Number of days to include data}
}
\description{
\code{mAPCA} Apply mAPCA to pollution data
}
\examples{
data(nycdat)
mAPCA(list(nycdat))
}

