% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matchfun}
\alias{matchfun}
\title{\code{matchfun} Matches major and local source signatures}
\usage{
matchfun(source.sig, major.sig, thres = pi/4)
}
\arguments{
\item{source.sig}{local source signatures from all monitors}

\item{major.sig}{major source signatures}

\item{thres}{cutoff for angle between local and major sources}
}
\description{
\code{matchfun} Matches major and local source signatures
}

