% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{percinc}
\alias{percinc}
\title{\code{percinc} Percent increase}
\usage{
percinc(beta, scale = 10)
}
\arguments{
\item{beta}{regression coefficients from poisson model}

\item{scale}{amount for percent increase}
}
\description{
\code{percinc} Percent increase
}
