% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{toenv}
\alias{toenv}
\title{Convert an object store connection object to an environment full of promises}
\usage{
toenv(x, base = "")
}
\arguments{
\item{x}{an object store, see \code{connect()}, \code{\link{setAs}}}

\item{base}{base path}
}
\value{
an environment
}
\description{
Convert an object store connection object to an environment full of promises
}
\examples{
\dontrun{
# Start an example local mongoose backend server
mongoose_start(path=tempdir())
con <- connect()
# Store some data as objects...
assign("cars", head(cars), con)
assign("Nile", head(Nile), con)
# ...more objects in a directory named 'mydata':
assign("mydata/iris", head(iris), con)
# Represent the connection as an environment
e = as(con, "environment")
ls(e)
e$cars
(e$"mydata/")$iris
mongoose_stop()
}
}
\keyword{function}
\keyword{internal}

