% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsources.R
\name{vPCA}
\alias{vPCA}
\title{\code{vPCA} applies PCA with a varimax rotation}
\usage{
vPCA(data, nsources = NULL, cut = 1)
}
\arguments{
\item{data}{data frame of daily constituent concentrations}

\item{cut}{cutoff for eigenvalues (see nsource), default is 1.}

\item{nsource}{number of sources.  If null, uses number of eigenvalues of the correlation matrix greater than cut.}
}
\description{
Applies PCA with a varimax rotation to data from a single monitor
}
